requirements.txt
PyYAML==6.0.2
markdown==3.6
pypandoc==1.13
python-docx==1.1.2
ebooklib==0.18
